refactor: move location button from generic search input to specific
