+++ src/features/search/components/SearchMainInput/SearchMainInput.tsx
@@ -2,8 +2,12 @@ import React, { FunctionComponent } from 'react'
 import { NativeSyntheticEvent, TextInputSubmitEditingEventData } from 'react-native'
 import styled from 'styled-components/native'
 
+import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
+import { styledButton } from 'ui/components/buttons/styledButton'
 import { SearchInput } from 'ui/components/inputs/SearchInput'
+import { LocationPointer } from 'ui/svg/icons/LocationPointer'
 import { MagnifyingGlass } from 'ui/svg/icons/MagnifyingGlass'
+import { getSpacing } from 'ui/theme'
 
 type QueryProps = {
   query?: string
@@ -51,15 +55,29 @@ export const SearchMainInput: FunctionComponent<Props> = ({
     autoFocus={isFocus}
     onFocusState={onFocusState}
     nativeAutoFocus
-    locationLabel={locationLabel}
-    onPressLocationButton={showLocationButton ? onPressLocationButton : undefined}
     LeftIcon={MagnifyingGlassIcon}
     inputHeight="regular"
     testID="searchInput"
-    {...props}
-  />
+    {...props}>
+    {!!showLocationButton && (
+      <LocationButton
+        testID="locationButton"
+        wording={locationLabel || ''}
+        onPress={onPressLocationButton}
+        icon={LocationPointer}
+        buttonHeight="extraSmall"
+        ellipsizeMode="tail"
+      />
+    )}
+  </SearchInput>
 )
 
 const MagnifyingGlassIcon = styled(MagnifyingGlass).attrs(({ theme }) => ({
   size: theme.icons.sizes.smaller,
 }))``
+
+const LocationButton = styledButton(ButtonPrimary)({
+  maxWidth: 142, // max width corresponds to the size of "Autour de moi" state.
+  marginLeft: getSpacing(2),
+  marginRight: -getSpacing(2),
+})
