+++ src/ui/components/inputs/SearchInput.stories.tsx
@@ -15,48 +15,34 @@ const Template: ComponentStory<typeof SearchInput> = (args) => <SearchInput {...
 export const Default = Template.bind({})
 Default.args = {
   placeholder: 'Placeholder...',
-  onPressLocationButton: undefined,
 }
 
 export const Disabled = Template.bind({})
 Disabled.args = {
   placeholder: 'Placeholder...',
   disabled: true,
-  onPressLocationButton: undefined,
 }
 
 export const WithTallHeight = Template.bind({})
 WithTallHeight.args = {
   placeholder: 'Placeholder...',
   inputHeight: 'tall',
-  onPressLocationButton: undefined,
 }
 
 export const WithLeftIcon = Template.bind({})
 WithLeftIcon.args = {
   placeholder: 'Placeholder...',
   LeftIcon: () => <MagnifyingGlass />,
-  onPressLocationButton: undefined,
 }
 
 export const WithLabel = Template.bind({})
 WithLabel.args = {
   placeholder: 'Placeholder...',
   label: 'Label',
-  onPressLocationButton: undefined,
 }
 
 export const WithValue = Template.bind({})
 WithValue.args = {
   placeholder: 'Placeholder...',
   value: 'Value',
-  onPressLocationButton: undefined,
-}
-
-export const WithLocationButton = Template.bind({})
-WithLocationButton.args = {
-  placeholder: 'Placeholder...',
-  value: '',
-  locationLabel: 'Autour de moi',
-  inputHeight: 'regular',
 }
